Extracting SfJ tests into a separate file
